js display cpu boards, beans
worked on displaying cpu boards works most of the time but could be that cpu1 and 2 select the same board. Worked on the bean making the dragable need to have them in object to place them on the boards. Made multiple functions. There is only two days left need to hurrry ASAP !!!!!!!!!!!